refactory: create translate adapter
